refactor: StationMangeApp이 가지고 있는 책임변경
StationMangeApp이 가지고 있는 책임을 StationMange가 가지게 함
